CleanUp, fix namespace, remove unused usage
