+++ MarketApp/MarketApp.csproj
@@ -20,8 +20,7 @@
 		<PackageReference Include="AdonisUI" Version="1.17.1" />
 		<PackageReference Include="AdonisUI.ClassicTheme" Version="1.17.1" />
 		<PackageReference Include="Meziantou.WpfFontAwesome" Version="5.15.4" />
-		<PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.9" />
-		<PackageReference Include="Telegram.Bot" Version="18.0.0" />
+		<PackageReference Include="Telegram.Bot" Version="19.0.0" />
 	</ItemGroup>
 
 	<ItemGroup>
