+++ MarketApp/Pages/SellPage.xaml
@@ -121,16 +121,7 @@
 
         <Border BorderThickness="1" BorderBrush="{DynamicResource {x:Static adonisUi:Brushes.Layer1HighlightBorderBrush}}"
         Grid.Column="1" Grid.Row="0">
-            <Border.Style>
-                <Style TargetType="Border">
-                    <Style.Triggers>
-                        <DataTrigger Binding="{Binding IsMouseOver, ElementName=ItemsLB}" Value="True">
-                            <Setter Property="ScrollViewer.VerticalScrollBarVisibility" Value="Auto" />
-                        </DataTrigger>
-                    </Style.Triggers>
-                </Style>
-            </Border.Style>
-
+          
             <ScrollViewer VerticalScrollBarVisibility="Auto">
                 <StackPanel>
                     <fa:FontAwesomeIcon x:Name="Spinner1" SolidIcon="Spinner" Margin="15" AnimationType="Spin" FontSize="192"
